+++ src/generators/Generator.js
@@ -9,11 +9,13 @@ import getOutro from './shared/utils/getOutro.js';
 import annotateWithScopes from './annotateWithScopes.js';
 
 export default class Generator {
-	constructor ( parsed, source, names, visitors ) {
+	constructor ( parsed, source, name, names, visitors, options ) {
 		this.parsed = parsed;
 		this.source = source;
+		this.name = name;
 		this.names = names;
 		this.visitors = visitors;
+		this.options = options;
 
 		this.imports = [];
 		this.helpers = {};
@@ -98,7 +100,7 @@ export default class Generator {
 
 						if ( globalWhitelist[ name ] ) {
 							code.prependRight( node.start, `( '${name}' in root ? root.` );
-							code.appendLeft( node.object.end, ` : ${name} )` );
+							code.appendLeft( node.object ? node.object.end : node.end, ` : ${name} )` );
 						} else {
 							code.prependRight( node.start, `root.` );
 						}
@@ -360,4 +362,3 @@ export default class Generator {
 		if ( visitor.leave ) visitor.leave( this, node );
 	}
 }
-