+++ src/generators/server-side-rendering/visitors/Component.js
@@ -41,13 +41,20 @@ export default {
 
 				return `${attribute.name}: ${value}`;
 			})
+			.concat( bindings.map( binding => {
+				const parts = binding.value.split( '.' );
+				const value = parts[0] in generator.current.contexts ? binding.value : `root.${binding.value}`;
+				return `${binding.name}: ${value}`;
+			}))
 			.join( ', ' );
 
+		const expression = node.name === ':Self' ? generator.name : `template.components.${node.name}`;
+
 		bindings.forEach( binding => {
-			generator.addBinding( binding, node.name );
+			generator.addBinding( binding, expression );
 		});
 
-		let open = `\${template.components.${node.name}.render({${props}}`;
+		let open = `\${${expression}.render({${props}}`;
 
 		if ( node.children.length ) {
 			open += `, { yield: () => \``;