+++ test/formats/index.js
@@ -1,6 +1,6 @@
-import deindent from '../src/utils/deindent.js';
+import deindent from '../../src/utils/deindent.js';
 import assert from 'assert';
-import { svelte, env, setupHtmlEqual } from './helpers.js';
+import { svelte, env, setupHtmlEqual } from '../helpers.js';
 
 function testAmd ( code, expectedId, dependencies, html ) {
 	const fn = new Function( 'define', code );
@@ -50,6 +50,21 @@ function testCjs ( code, dependencyById, html ) {
 
 function testIife ( code, name, globals, html ) {
 	const fn = new Function( Object.keys( globals ), `${code}\n\nreturn ${name};` );
+
+	return env().then( window => {
+		const SvelteComponent = fn( ...Object.keys( globals ).map( key => globals[ key ] ) );
+
+		const main = window.document.body.querySelector( 'main' );
+		const component = new SvelteComponent({ target: main });
+
+		assert.htmlEqual( main.innerHTML, html );
+
+		component.destroy();
+	});
+}
+
+function testEval ( code, name, globals, html ) {
+	const fn = new Function( Object.keys( globals ), `return ${code};` );
 
 	return env().then( window => {
 		const SvelteComponent = fn( ...Object.keys( globals ).map( key => globals[ key ] ) );
@@ -173,6 +188,33 @@ describe( 'formats', () => {
 			return testAmd( code, 'foo', { answer: 42 }, `<div>42</div>` )
 				.then( () => testCjs( code, { answer: 42 }, `<div>42</div>` ) )
 				.then( () => testIife( code, 'Foo', { answer: 42 }, `<div>42</div>` ) );
+		});
+	});
+
+	describe( 'eval', () => {
+		it( 'generates a self-executing script that returns the component on eval', () => {
+			const source = deindent`
+				<div>{{answer}}</div>
+
+				<script>
+					import answer from 'answer';
+
+					export default {
+						data () {
+							return { answer };
+						}
+					};
+				</script>
+			`;
+
+			const { code } = svelte.compile( source, {
+				format: 'eval',
+				globals: {
+					answer: 'answer'
+				}
+			});
+
+			return testEval( code, 'Foo', { answer: 42 }, `<div>42</div>` );
 		});
 	});
 });